cc_requirements.txt
torch==1.12
torch_geometric==2.2
torch_sparse==0.6.14
torch_scatter==2.0.9
tensorboard==2.10.0
scikit-learn==1.1
tqdm==4.64.1
networkx==2.5.1
pyyaml==6.0
pillow==9.2.0
h5py==3.6.0
lxml==4.9.1
pyproj==3.3.0
folium==0.12.1.post1
geojson==2.5.0
PyMySQL==0.9.3
optuna==3.0.5
sqlalchemy==1.4.31
hydra-core==1.2.0
